Outout Byte Order Mark iff first argument has one
If the contents of the first file argument start with a Unicode Byte
Order Mark, then if we have any lines to output we begin the output with
a UTF-8 Byte Order Mark.

If a UTF-16LE and UTF-16BE file starts with a Byte Order Mark, it will
be translated to UTF8 for processing.