If `line` is in a CountedSet, don't allocate a key
For the second and subsequent file arguments of the `Single` and
`Multiple` operations, we create a slice set of the file's lines,
then set the `CountedSet`'s value for that line to `FoundIn::One` if it
wasn't already present, or to `FoundIn::Many` if it was. In the first
case we always need to allocate a key so it will survive dropping the
file contents. But if we've found a value with `set.get_mut()`, we
can just set it to FoundIn::Many.