Move `for_byte_lines` from `NextOperand` to a trait
`NextOperand` implicitly uses file I/O operations, which are inefficient
for property testing and fuzzing.  But the `calculate` function only
uses the `for_byte_lines` method of the items in its `rest` iterator
parameter, so we extract `for_byte_lines` into the `LaterOperand` trait.

Currently we `impl LaterOperand for &[u8]` and use that in tests only as
a check on the write-to-file version with `NextOperand`. In the future
we can write tests to check that `operations::calculate` handles errors
correctly, and write property based tests that aren't dreadfully slow.