Refactor: provide replacements for .get_mut
In a `CountedSet`s, the `IndexMap` value will have a count field
we don't want clients to see; so move the code that uses `.get_mut` into
the `ZetSet` trait. This couples us more tightly to the algorithms in
the `operations` module, but I don't see how to get around it at this
point.