Abandon most of the last seven commits
Instead of making ZetSet a trait with CountedSet and UncountedSet
implementations, we're going to adjust the Bookkeeping trait so
that B: Bookkeeping can have both counted and uncounted variants.

We keep a bug fix in main.rs, and the tests that check for it — though the
tests are currently disabled until the code that revealed bug is
recreated! (See be031ccb5295ffd7d542bd5af9c00fac1b611c81)